Remove useless header file
